+++ src/boot/ksdk1.1.0/devMAG3110.h
@@ -41,7 +41,7 @@
 
 void		initMAG3110(const uint8_t i2cAddress, WarpI2CDeviceState volatile *  deviceStatePointer);
 WarpStatus	writeSensorRegisterMAG3110(uint8_t deviceRegister, uint8_t payload, uint16_t menuI2cPullupValue);
-WarpStatus	configureSensorMAG3110(uint8_t payloadCTRL_REG1, uint8_t payloadCTRL_REG2, uint8_t menuI2cPullupValue);
+WarpStatus	configureSensorMAG3110(uint8_t payloadCTRL_REG1, uint8_t payloadCTRL_REG2, uint16_t menuI2cPullupValue);
 WarpStatus	readSensorRegisterMAG3110(uint8_t deviceRegister, int numberOfBytes);
 WarpStatus	readSensorSignalMAG3110(WarpTypeMask signal,
 					WarpSignalPrecision precision,
